Fix frontend lint findings: rules-of-hooks, setState-in-effect, refs, fast-refresh (#141)
Fixes the real bugs from the lint audit: Footer.jsx's conditional useState
call (rules-of-hooks), a cluster of setState-in-effect violations across
filter/table/map components (resolved by deriving local UI state during
render per React's documented pattern instead of syncing via effect, or by
nesting async fetch logic one function deeper so the effect body itself
never calls setState directly), two ref-safety violations (a shared
module-level calendar-month variable lifted into ResortDetailModal's own
state; ResortMap's hover tooltip reading containerRef during render, now
tracked via a ResizeObserver-backed state value instead), an unused `error`
var, and vite.config.js's CJS __dirname under ESM. Also splits ResortTable's
COLUMN_DEFS/HEADER_BY_FIELD/COL_GROUPS and BoolCell/LinkCell cell renderers
into their own files so ResortTable.jsx and Panel.jsx satisfy
react-refresh/only-export-components (component-only exports).

Down from 16 errors/14 warnings to 0 errors/4 warnings — the remaining
warnings are pre-existing, deliberate dependency-array omissions (e.g.
ResortMap's zoom-fit effect intentionally excludes `filters` per its
existing comment) left as-is to avoid behavioral regressions.
