+++ CLAUDE.md
@@ -32,7 +32,7 @@ The aesthetic direction, dark mode token config, glow effects, and open design q
 
 - **`src/theme.js`** — Single source of truth for all colors (`COLORS`), fonts (`FONTS`), map dot colors (`MAP_DOT_COLORS`), Ant Design theme config, and the `withAlpha(hex, alpha)` utility. **All color values in `.jsx` and `.css` files must reference named tokens from `COLORS` — no hardcoded hex or rgba strings anywhere in `src/`.**
 - **`src/App.jsx`** — Top-level layout. Owns table state: `gridRef` (forwarded to AG Grid), `tablePanelRef` (fullscreen target), column visibility, and toolbar button handlers. The drag handle row contains the collapse toggle (left) and table controls (right): Select Columns popover, Download CSV, Full screen.
-- **`src/components/ResortTable.jsx`** — AG Grid wrapper. A `forwardRef` component that forwards `ref` to `<AgGridReact>`. Exports `COLUMN_DEFS`, `HEADER_BY_FIELD`, `COL_GROUPS` for use by App.jsx's column-visibility popover. No internal toolbar — all controls live in App.jsx's drag handle.
+- **`src/components/ResortTable.jsx`** — AG Grid wrapper. A `forwardRef` component that forwards `ref` to `<AgGridReact>`. No internal toolbar — all controls live in App.jsx's drag handle. Column data lives in `src/components/resortTableColumns.jsx` (`COLUMN_DEFS`, `HEADER_BY_FIELD`, `COL_GROUPS`, used by App.jsx's column-visibility popover), with `BoolCell`/`LinkCell` cell renderers split into `src/components/resortTableCells.jsx` — kept out of ResortTable.jsx itself so the file only exports the component, per `react-refresh/only-export-components`.
 - **`src/components/ResortTable.css`** — AG Grid header overrides only. Colors are set as CSS custom properties via `GRID_THEME_VARS` in ResortTable.jsx (e.g. `--indy-header-bg`) so the CSS file stays free of hardcoded values.
 
 ### Key Design Patterns (Pipeline)
