+++ frontend/src/components/common/panelStyle.js
@@ -1,0 +1,7 @@
+import { COLORS } from '@/theme'
+
+export const PANEL_STYLE = {
+  background: COLORS.bgBase,
+  border: `2px solid ${COLORS.bgHeader}`,
+  borderRadius: 4,
+}
