+++ frontend/src/components/filters/PeakRankingsFilters.jsx
@@ -17,9 +17,13 @@ function RangeSlider({ label, metaRange, minKey, maxKey }) {
 
   const [local, setLocal] = useState([urlMin, urlMax])
 
-  useEffect(() => {
-    setLocal([filters[minKey] ?? metaMin, filters[maxKey] ?? metaMax])
-  }, [filters[minKey], filters[maxKey], metaMin, metaMax])
+  // Reset local state when the URL value changes externally — adjusted during render
+  // rather than via effect, per https://react.dev/learn/you-might-not-need-an-effect
+  const [syncedRange, setSyncedRange] = useState([urlMin, urlMax])
+  if (urlMin !== syncedRange[0] || urlMax !== syncedRange[1]) {
+    setSyncedRange([urlMin, urlMax])
+    setLocal([urlMin, urlMax])
+  }
 
   function onChangeComplete([min, max]) {
     setFilters({
@@ -64,10 +68,18 @@ function CategoricalSelect({ label, filterKey, options }) {
   const [local, setLocal] = useState(urlValue)
   const ref = useRef(urlValue)
 
-  useEffect(() => {
+  // Reset local state when the URL value changes externally — adjusted during render
+  // rather than via effect, per https://react.dev/learn/you-might-not-need-an-effect
+  const urlKey = JSON.stringify(urlValue)
+  const [syncedKey, setSyncedKey] = useState(urlKey)
+  if (urlKey !== syncedKey) {
+    setSyncedKey(urlKey)
     setLocal(urlValue)
+  }
+  // Ref writes aren't allowed during render, so this stays in an effect
+  useEffect(() => {
     ref.current = urlValue
-  }, [JSON.stringify(urlValue)])
+  }, [urlKey, urlValue])
 
   function onChange(v) {
     setLocal(v)
