+++ frontend/vite.config.js
@@ -1,7 +1,10 @@
 import path from 'path'
+import { fileURLToPath } from 'url'
 import { defineConfig } from 'vitest/config'
 import react from '@vitejs/plugin-react'
 import tailwindcss from '@tailwindcss/vite'
+
+const __dirname = path.dirname(fileURLToPath(import.meta.url))
 
 // https://vite.dev/config/
 export default defineConfig({
